requirements/base.txt
python-slugify==8.0.4  # https://github.com/un33k/python-slugify
Pillow==10.3.0  # https://github.com/python-pillow/Pillow
argon2-cffi==23.1.0  # https://github.com/hynek/argon2_cffi
whitenoise==6.7.0  # https://github.com/evansd/whitenoise
redis==5.0.6  # https://github.com/redis/redis-py
hiredis==2.3.2  # https://github.com/redis/hiredis-py
uvicorn[standard]==0.30.1  # https://github.com/encode/uvicorn
uvicorn-worker==0.2.0  # https://github.com/Kludex/uvicorn-worker

# Django
# ------------------------------------------------------------------------------
django==4.2.13  # pyup: < 5.0  # https://www.djangoproject.com/
django-environ==0.11.2  # https://github.com/joke2k/django-environ
django-model-utils==4.5.1  # https://github.com/jazzband/django-model-utils
django-allauth[mfa]==0.63.3  # https://github.com/pennersr/django-allauth
django-crispy-forms==2.2  # https://github.com/django-crispy-forms/django-crispy-forms
crispy-bootstrap5==2024.2  # https://github.com/django-crispy-forms/crispy-bootstrap5
django-redis==5.4.0  # https://github.com/jazzband/django-redis


django-auto-prefetch==1.4.0
uuid==1.30

shortuuid==1.0.13  #https://github.com/stochastic-technologies/shortuuid

django-resized==1.0.2 # https://github.com/un1t/django-resized/blob/master/django_resized/forms.py

django-multiupload==0.6.1

# django-ckeditor-5==0.2.13
django-ckeditor==6.7.1

django-countries==7.6.1

django-taggit==5.0.1

cloudinary==1.40.0  #https://cloudinary.com/documentation/django_image_and_video_upload#django_forms_and_models
requests==2.32.3  #https://requests.readthedocs.io/en/latest/

django-jazzmin==3.0.0  #https://django-jazzmin.readthedocs.io/

nowpayments==1.3.3
